applying facade
