+++ src/router/routes.js
@@ -3,12 +3,12 @@ import Router from 'vue-router'
 import Home from '../views/LandingPage/Home.vue'
 import OrderBagPage from '../views/OrderBag/orderBagPage.vue'
 import ShoppingPage from '../views/ShoppingPage/ShoppingPage.vue'
-import CategoriePage from '../views/CategoriePage.vue'
-import EditUser from '../views/EditUserProfile/EditUser.vue'
+import CategoryPage from '../views/CategoryPage.vue'
+import EditUserProfile from '../views/EditUserProfile.vue'
 import LoggedUserHeader from '../components/LoggedUserHeader.vue'
-import Menu from '../views/MenuPage/Menu.vue'
+import RestaurantMenu from '../views/MenuPage/RestaurantMenu.vue'
 import Item from '../views/ItemPage/Item.vue'
-import UserPage from '../views/UserProfile/Data.vue'
+import UserProfile from '../views/UserProfile.vue'
 
 Vue.use(Router)
 
@@ -29,45 +29,45 @@ export default new Router({
             }
         },
         {
-            path: '/categorie/:name',
+            path: '/categoria/:name',
             name: 'categoria',
             components: {
-                default: CategoriePage
+                default: CategoryPage
             }
         },
         {
-            path: '/edit-user',
+            path: '/editar-usuario',
             name: 'edit-user',
             components: {
-                default: EditUser,
+                default: EditUserProfile,
                 LoggedUserHeader
 
             }
         },
         {
-            path: '/user',
+            path: '/usuario',
             name: 'user',
             components: {
-                default: UserPage,
+                default: UserProfile,
                 LoggedUserHeader
             }
         },
         {
-            path: '/menu-page',
+            path: '/restaurante/:cnpj',
             name: 'menu',
             components: {
-                default: Menu
+                default: RestaurantMenu
             }
         },
         {
-            path: '/item-page',
+            path: '/item/:id',
             name: 'item',
             components: {
                 default: Item
             }
         },
         {
-            path: '/orderbag',
+            path: '/sacola',
             name: 'order-bag',
             components: {
                 default: OrderBagPage,
