+++ src/services/ServicesFacade.js
@@ -1,4 +1,4 @@
-import { getAllRestaurants, getRestaurant} from './restaurantService'
+import { getAllRestaurants, getRestaurant, getRestaurantsByCategory, getRestaurantMenu, getRestaurantCategories} from './restaurantService'
 import { getShopping } from './shoppingService'
 import { getUser } from './userService'
 
@@ -6,5 +6,8 @@ export default {
     getShopping,
     getAllRestaurants,
     getRestaurant,
-    getUser
+    getUser,
+    getRestaurantsByCategory,
+    getRestaurantMenu,
+    getRestaurantCategories
 }
