--- src/views/UserProfile/User.vue
@@ -1,26 +1,0 @@
-<template>
-  <v-content class="d-flex flex-column pt-0">
-    <div>
-
-      <QrcData />
-    </div>
-  </v-content>
-</template>
-
-<script>
-import Banner from "./Banner.vue";
-import Data from "./Data.vue";
-export default {
-  components: {
-    QrcBanner: Banner,
-    QrcData: Data
-  }
-};
-</script>
-
-<style lang="scss">
-.v-content__wrap {
-  display: flex;
-  flex-direction: column;
-}
-</style>
